pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "mammadov115",email = "mehman.mammadov182@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django == 5.2.4",
    "djangorestframework == 3.16.0",
    "dotenv == 0.9.9",
    "requests == 2.32.4",
    "psycopg2-binary == 2.9.10",
    "djangorestframework-simplejwt == 5.5.1",
    "drf-yasg == 1.21.10",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "django-unfold (>=0.76.0,<0.77.0)",
]

[tool.poetry]
packages = [{include = "project", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"
pytest-mock = "^3.14.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
